requirements.txt
timm
numpy
pillow
scipy
transformers
accelerate
datasets
sentencepiece
tokenizers
nltk
rouge-score
torchmetrics[multimodal]
lpips
scikit-image
matplotlib
tqdm
pyyaml
wandb
